Item checking and removal!
In Option.java:

changed constructor method in Option to include item needed, and text for if you don't have the item.

Fixed weird error where ArrayList<Item>.contains(Item foo) doesn't work. Same with .remove(Item bar)

In Ship.java:

updated option constructors for Option

removed faux if statement for setting Option based on user having glass.

added getReadyRoom() method, returns the ready room so TextAdventure can check if the user is in that room.

In TextAdventure.java:

put if statement into while loop. It constantly checks to see if the user is in the ready room, and changes the Option5 to "put glass on table" if they have a glass. While loop's probably not the most elegent way to do this, but it works.